# improve hexDump performance