@christoph2 2020-11-11
Thanks Daniel, 

BTW:  for some time now, I'm working on a "secret" / unpublished project focusing on integration and high-level API aspects (including asammdf,  of course). Hopefully I can publish an initial release within the next few days.