@danielhrisca 2020-11-11
Sounds great

I'm also developing a project at work that is very similar to CANape for data acquisition and offline visualization (no calibration for the moment). I don't think I will be able to publish that, but at least I can give some advice if necessary